pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finanbot"
dynamic = ["version"]
description = "Personal finance tracker with FastAPI and Streamlit and some bot help"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = []
authors = [
  { name = "Anderson R Damasceno", email = "anderdam@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = [
	"alembic>=1.17.0",
	"asyncpg>=0.30.0",
	"bandit>=1.8.6",
	"black>=25.9.0",
	"commitizen>=4.9.1",
	"fastapi>=0.119.0",
	"httpx>=0.28.1",
	"isort>=7.0.0",
	"pandas>=2.3.3",
	"passlib>=1.7.4",
	"pre-commit>=4.3.0",
	"psycopg2-binary>=2.9.11",
	"pydantic-settings>=2.11.0",
	"pydantic>=2.12.2",
    "pydantic[email]",
	"pytest-asyncio>=1.2.0",
	"pytest>=8.4.2",
	"python-dotenv>=1.1.1",
	"python-jose>=3.5.0",
	"python-multipart>=0.0.20",
	"rclone_python>=0.1.23",
	"requests>=2.32.5",
	"richer>=0.1.6",
	"ruff>=0.14.0",
	"sqlalchemy>=2.0.44",
	"streamlit>=1.50.0",
	"uvicorn>=0.37.0"
]

[project.urls]
Documentation = "https://github.com/anderdam/finanbot#readme"
Issues = "https://github.com/anderdam/finanbot/issues"
Source = "https://github.com/anderdam/finanbot"

[tool.hatch.version]
path = "src/__about__.py"

[tool.hatch.envs.dev.scripts]
pcr = "pre-commit run --all-files"
startapp = "uvicorn src.main:app --reload"
startui = "streamlit run ui/streamlit_app.py"
aleup = "alembic upgrade head"
lint = "ruff src tests"
format = "black src tests"
test = "pytest"

[tool.coverage.run]
source_pkgs = ["src", "tests"]
branch = true
parallel = true
omit = ["src/__about__.py"]

[tool.coverage.paths]
finanbot = ["src", "*/finanbot/src"]
tests = ["tests", "*/finanbot/tests"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "B"]
lint.ignore = ["B008"]
exclude = ["tests"]
target-version = "py313"

[tool.isort]
profile = "black"

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pyright]
include = ["src"]
exclude = [
  "**/node_modules",
  "**/__pycache__",
  "src/experimental",
  "src/typestubs"
]
ignore = ["src/oldstuff"]
defineConstant = { DEBUG = true }
stubPath = "src/stubs"

reportMissingImports = false
reportMissingTypeStubs = false

pythonVersion = "3.13"
pythonPlatform = "Windows"

executionEnvironments = [
  { root = "src", pythonVersion = "3.13", pythonPlatform = "Windows", extraPaths = [ "src/service_libs" ], reportMissingImports = "warning" },
  { root = "tests", extraPaths = ["tests/"]},
  { root = "src" }
]
